separated out highlighting from Cytoscape.svelte
